Extract CommitShaProviders into separate directory
