+++ affectedmoduledetector/src/main/kotlin/com/dropbox/affectedmoduledetector/commitshaproviders/CommitShaProvider.kt
@@ -1,0 +1,18 @@
+package com.dropbox.affectedmoduledetector.commitshaproviders
+
+import com.dropbox.affectedmoduledetector.GitClient
+import com.dropbox.affectedmoduledetector.Sha
+
+interface CommitShaProvider {
+    fun getCommitSha(commandRunner: GitClient.CommandRunner): Sha?
+
+    companion object {
+        fun fromString(string: String): CommitShaProvider {
+            return when (string) {
+                "PreviousCommit" -> PreviousCommit()
+                else -> ForkCommit()
+            }
+        }
+    }
+}
+
