+++ affectedmoduledetector/src/main/kotlin/com/dropbox/affectedmoduledetector/commitshaproviders/PreviousCommit.kt
@@ -1,0 +1,16 @@
+package com.dropbox.affectedmoduledetector.commitshaproviders
+
+import com.dropbox.affectedmoduledetector.GitClient
+import com.dropbox.affectedmoduledetector.Sha
+
+class PreviousCommit: CommitShaProvider {
+    override fun getCommitSha(commandRunner: GitClient.CommandRunner): Sha? {
+        return commandRunner.executeAndParse(PREV_COMMIT_CMD)
+            .firstOrNull()
+            ?.split(" ")
+            ?.firstOrNull()
+    }
+    companion object {
+        const val PREV_COMMIT_CMD = "git --no-pager rev-parse HEAD~1"
+    }
+}
